02_streamlit_chatgpt/requirements.txt
openai
python-decouple
streamlit
streamlit-chat
langchain
